+++ src/DependencyGraph/DependencyGraph.ts
@@ -6,12 +6,7 @@
 import assert from 'assert'
 import {AbsoluteCellRange} from '../AbsoluteCellRange'
 import {absolutizeDependencies} from '../absolutizeDependencies'
-import {
-  CellError,
-  ErrorType,
-  simpleCellAddress,
-  SimpleCellAddress
-} from '../Cell'
+import {CellError, ErrorType, simpleCellAddress, SimpleCellAddress} from '../Cell'
 import {RawCellContent} from '../CellContentParser'
 import {CellDependency} from '../CellDependency'
 import {Config} from '../Config'
@@ -39,15 +34,17 @@ import {Graph, TopSortResult} from './Graph'
 import {MatrixMapping} from './MatrixMapping'
 import {RangeMapping} from './RangeMapping'
 import {SheetMapping} from './SheetMapping'
-import {RawAndParsedValue, ValueCellVertexValue} from './ValueCellVertex'
+import {RawAndParsedValue} from './ValueCellVertex'
 import {FunctionRegistry} from '../interpreter/FunctionRegistry'
 import {
   EmptyValue,
+  ExtendedNumber,
   getRawValue,
   InternalScalarValue,
   InterpreterValue,
-  RawScalarValue, ExtendedNumber
+  RawScalarValue
 } from '../interpreter/InterpreterValue'
+import {MatrixSize} from '../Matrix'
 
 export class DependencyGraph {
   /*
@@ -154,6 +151,16 @@ export class DependencyGraph {
       this.removeVertex(vertex)
       this.addressMapping.removeCell(address)
     }
+  }
+
+  public setMatrixToCell(address: SimpleCellAddress, ast: Ast, dependencies: CellDependency[], size: MatrixSize) {
+    const vertex = new MatrixVertex(address, size.width, size.height, ast)
+    this.exchangeOrAddMatrixVertex(vertex)
+
+    this.processCellDependencies(dependencies, vertex)
+
+    this.graph.markNodeAsSpecialRecentlyChanged(vertex)
+    this.correctInfiniteRangesDependency(address)
   }
 
   public ensureThatVertexIsNonMatrixCellVertex(vertex: CellVertex | null) {
@@ -548,30 +555,12 @@ export class DependencyGraph {
 
   public addVertex(address: SimpleCellAddress, vertex: CellVertex): void {
     this.graph.addNode(vertex)
-    this.setVertexAddress(address, vertex)
+    this.addressMapping.setCell(address, vertex)
   }
 
   public addMatrixVertex(address: SimpleCellAddress, vertex: CellVertex): void {
     this.graph.addNode(vertex)
     this.setAddressMappingForMatrixVertex(vertex, address)
-  }
-
-  public addNewMatrixVertex(matrixVertex: MatrixVertex): void {
-    const range = AbsoluteCellRange.spanFrom(matrixVertex.getAddress(), matrixVertex.width, matrixVertex.height)
-    for (const vertex of this.verticesFromRange(range)) {
-      if (vertex instanceof MatrixVertex) {
-        throw Error('You cannot modify only part of an array')
-      }
-    }
-
-    this.setMatrix(range, matrixVertex)
-
-    for (const [address, vertex] of this.entriesFromRange(range)) {
-      if (vertex) {
-        this.exchangeGraphNode(vertex, matrixVertex)
-      }
-      this.setVertexAddress(address, matrixVertex)
-    }
   }
 
   public* matrixFormulaNodes(): IterableIterator<MatrixVertex> {
@@ -616,10 +605,6 @@ export class DependencyGraph {
       return new CellError(ErrorType.VALUE, ErrorMessage.ScalarExpected)
     }
     return value
-  }
-
-  public setVertexAddress(address: SimpleCellAddress, vertex: CellVertex) {
-    this.addressMapping.setCell(address, vertex)
   }
 
   public existsEdge(fromNode: Vertex, toNode: Vertex): boolean {
@@ -881,8 +866,22 @@ export class DependencyGraph {
     }
   }
 
+  private exchangeOrAddMatrixVertex(matrixVertex: MatrixVertex): void {
+    const range = AbsoluteCellRange.spanFrom(matrixVertex.getAddress(), matrixVertex.width, matrixVertex.height)
+    for (const vertex of this.verticesFromRange(range)) {
+       this.ensureThatVertexIsNonMatrixCellVertex(vertex)
+    }
+
+    this.setMatrix(range, matrixVertex)
+
+    for (const [address, vertex] of this.entriesFromRange(range)) {
+      this.exchangeOrAddGraphNode(vertex, matrixVertex)
+      this.addressMapping.setCell(address, matrixVertex)
+    }
+  }
+
   private setAddressMappingForMatrixVertex(vertex: CellVertex, formulaAddress: SimpleCellAddress): void {
-    this.setVertexAddress(formulaAddress, vertex)
+    this.addressMapping.setCell(formulaAddress, vertex)
 
     if (!(vertex instanceof MatrixVertex)) {
       return
@@ -892,7 +891,7 @@ export class DependencyGraph {
     this.setMatrix(range, vertex)
 
     for (const address of range.addresses(this)) {
-      this.setVertexAddress(address, vertex)
+      this.addressMapping.setCell(address, vertex)
     }
   }