+++ src/interpreter/plugin/MatrixPlugin.ts
@@ -8,9 +8,9 @@ import {ErrorMessage} from '../../error-message'
 import {MatrixSize, matrixSizeForMultiplication, matrixSizeForPoolFunction, matrixSizeForTranspose} from '../../Matrix'
 import {ProcedureAst} from '../../parser'
 import {Interpreter} from '../Interpreter'
-import {InterpreterState} from '../InterpreterState'
 import {InterpreterValue} from '../InterpreterValue'
 import {SimpleRangeValue} from '../SimpleRangeValue'
+import {InterpreterState} from '../InterpreterState'
 import {ArgumentTypes, FunctionPlugin, FunctionPluginTypecheck} from './FunctionPlugin'
 
 export type KernelRunShortcut = (...args: any[]) => number[][]
@@ -76,6 +76,9 @@ export class MatrixPlugin extends FunctionPlugin implements FunctionPluginTypech
     return this.runMatrixFunction(ast.args, state, this.metadata('MMULT'), (leftMatrix: SimpleRangeValue, rightMatrix: SimpleRangeValue) => {
       if (!leftMatrix.hasOnlyNumbers() || !rightMatrix.hasOnlyNumbers()) {
         return new CellError(ErrorType.VALUE, ErrorMessage.NumberRange)
+      }
+      if( rightMatrix.height() !== leftMatrix.width()) {
+        return new CellError(ErrorType.VALUE, ErrorMessage.MatrixDimensions)
       }
       const outputSize = matrixSizeForMultiplication(leftMatrix.size, rightMatrix.size)